# services/projects-service/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Projects Service API
  version: 1.0.0
  description: REST API for managing projects, members, tenants, and environments.
servers:
  - url: /
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Environment:
      type: object
      required: [id, project_id, name]
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
paths:
...
  /v1/projects/{projectId}/environments:
    get:
      summary: List environments
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Environment' } } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      description: Requires scope environments:read.
    post:
      summary: Create environment
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                config: { type: object, additionalProperties: true }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Environment' } } } }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      description: Requires scope environments:write.
...
  /v1/projects/{projectId}/environments/{envId}:
    put:
      summary: Update environment
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: path
          name: envId
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                config: { type: object, additionalProperties: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Environment' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      description: Requires scope environments:write.
    delete:
      summary: Delete environment
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: path
          name: envId
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      responses:
        '204': { description: No Content }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      description: Requires scope environments:write.
  /v1/projects/{projectId}/ingestion-modes:
    post:
      summary: Set ingestion modes
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modes, defaultMode]
              properties:
                modes: { type: array, items: { type: string } }
                defaultMode: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } }, required: [ok] } } } }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      description: Requires scope environments:write.